+++ src/ActorApi.Api/ActorApi.Api.csproj
@@ -4,6 +4,8 @@
     <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
+    <RootNamespace>ActorApi.Api</RootNamespace>
+    <AssemblyName>ActorApi.Api</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
