+++ src/ActorApi.Api/ActorApi.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35327.3
+# Visual Studio Version 18
+VisualStudioVersion = 18.6.11822.322
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActorApiDI", "ActorApiDI.csproj", "{D3D77F97-1AE8-4D0C-A14C-ADAB1E86B9CF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActorApi.Api", "ActorApi.Api.csproj", "{D3D77F97-1AE8-4D0C-A14C-ADAB1E86B9CF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTest", "..\UnitTest\UnitTest.csproj", "{1D42C063-08B5-4BE7-AB57-05A24EDE01C0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActorApi.Tests", "..\..\tests\ActorApi.Tests\ActorApi.Tests.csproj", "{5C534603-04D9-F48D-93CC-4FFF676C0576}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{D3D77F97-1AE8-4D0C-A14C-ADAB1E86B9CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D3D77F97-1AE8-4D0C-A14C-ADAB1E86B9CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D3D77F97-1AE8-4D0C-A14C-ADAB1E86B9CF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1D42C063-08B5-4BE7-AB57-05A24EDE01C0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D42C063-08B5-4BE7-AB57-05A24EDE01C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D42C063-08B5-4BE7-AB57-05A24EDE01C0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D42C063-08B5-4BE7-AB57-05A24EDE01C0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5C534603-04D9-F48D-93CC-4FFF676C0576}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5C534603-04D9-F48D-93CC-4FFF676C0576}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5C534603-04D9-F48D-93CC-4FFF676C0576}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5C534603-04D9-F48D-93CC-4FFF676C0576}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
